Split interpreter.cpp into many specialized files
